rename module to avoid conflict with python keyword
