+++ backend/ai_librarian_monorepo/ai_librarian_core/pyproject.toml
@@ -18,10 +18,10 @@ dependencies = [
     "playwright>=1.52.0",
     "pydantic>=2.11.5",
     "pyowm>=3.3.0",
-    "rich>=14.0.0",
     "setuptools>=80.9.0", # PyOWM requires setuptools
     "wikipedia>=1.4.0",
     "youtube-search>=2.1.2",
+    "rich>=14.1.0",
 ]
 
 [build-system]
