+++ backend/ai_librarian_monorepo/ai_librarian_core/src/ai_librarian_core/agents/react/synchronous.py
@@ -1,4 +1,4 @@
-from collections.abc import AsyncIterator
+from collections.abc import Iterator
 from dataclasses import dataclass
 from typing import Literal
 
@@ -7,14 +7,14 @@ from ai_librarian_core.agents.react.state import MessagesState
 from ai_librarian_core.agents.utils import get_thread_id
 from ai_librarian_core.models.llm_config import LLMConfig
 from ai_librarian_core.models.used_tool import UsedTool
-from langchain_core.messages import AIMessage, AnyMessage, BaseMessage, HumanMessage, ToolMessage
+from langchain_core.messages import AIMessage, BaseMessage, ToolMessage
 from langgraph.graph import StateGraph
 from langgraph.graph.state import CompiledStateGraph
 from langgraph.prebuilt import ToolNode
 
 
 @dataclass
-class AsyncReactAgent(BaseReactAgent):
+class ReactAgent(BaseReactAgent):
     def __post_init__(self):
         super().__post_init__()
 
@@ -22,22 +22,22 @@ class AsyncReactAgent(BaseReactAgent):
     def workflow(self) -> CompiledStateGraph:
         return self._init_workflow()
 
-    async def _clear_used_tools(self, state: MessagesState) -> dict[str, list[UsedTool]]:
+    def _clear_used_tools(self, state: MessagesState) -> dict[str, list[UsedTool]]:
         return {"used_tools": []}
 
-    async def _invoke_llm(self, state: MessagesState) -> dict[str, list[BaseMessage]]:
+    def _invoke_llm(self, state: MessagesState) -> dict[str, list[BaseMessage]]:
         llm_config = state.llm_config
         llm = self._init_llm(llm_config)
         messages = state.messages
 
         try:
-            response = await llm.ainvoke(messages)
+            response = llm.invoke(messages)
             return {"messages": [response]}
         # TODO(youkwan): Handle specific errors (couldn't find docs).
         except Exception as e:
             raise ReactAgentError("An unexpected error occurred while trying to invoke the chat model.") from e
 
-    async def _route(self, state: MessagesState) -> Literal["tools", "__end__"]:
+    def _route(self, state: MessagesState) -> Literal["tools", "__end__"]:
         messages = state.messages
         last_message = messages[-1]
         if not isinstance(last_message, AIMessage):
@@ -46,7 +46,7 @@ class AsyncReactAgent(BaseReactAgent):
             return "tools"
         return "__end__"
 
-    async def _catch_tool_massage(self, state: MessagesState) -> dict[str, list[UsedTool]]:
+    def _catch_tool_massage(self, state: MessagesState) -> dict[str, list[UsedTool]]:
         messages = state.messages
         used_tools = [
             UsedTool(name=msg.name, output=msg.content) for msg in reversed(messages) if isinstance(msg, ToolMessage)
@@ -74,47 +74,19 @@ class AsyncReactAgent(BaseReactAgent):
         workflow.add_edge("catch_tool_massage", "invoke_llm")
         return workflow.compile(name=self.name, checkpointer=self.checkpointer)
 
-    async def run(
-        self, messages: list[AnyMessage], thread_id: str | None = None, llm_config: LLMConfig = LLMConfig()
+    def run(
+        self, messages: list[BaseMessage], thread_id: str | None = None, llm_config: LLMConfig = LLMConfig()
     ) -> tuple[AIMessage, list[UsedTool]]:
         state = MessagesState(messages=messages, llm_config=llm_config)
-        result = await self.workflow.ainvoke(
-            state, config={"configurable": {"thread_id": thread_id or get_thread_id()}}
-        )
+        result = self.workflow.invoke(state, config={"configurable": {"thread_id": thread_id or get_thread_id()}})
         return result["messages"][-1], result["used_tools"]
 
-    async def stream(
-        self, messages: list[AnyMessage], thread_id: str | None = None, llm_config: LLMConfig = LLMConfig()
-    ) -> AsyncIterator[dict[str, list[BaseMessage]]]:
+    def stream(
+        self, messages: list[BaseMessage], thread_id: str | None = None, llm_config: LLMConfig = LLMConfig()
+    ) -> Iterator[dict[str, list[BaseMessage]]]:
         state = MessagesState(messages=messages, llm_config=llm_config)
-        return self.workflow.astream(
+        return self.workflow.stream(
             state,
             stream_mode="messages",
             config={"configurable": {"thread_id": thread_id or get_thread_id()}},
         )
-
-
-if __name__ == "__main__":
-    import asyncio
-
-    import dotenv
-
-    dotenv.load_dotenv()
-
-    from ai_librarian_core.tools.tools import init_built_in_tools
-
-    tools = init_built_in_tools()
-
-    agent = AsyncReactAgent(tools=tools)
-    messages, used_tools = asyncio.run(
-        agent.run([HumanMessage(content="請問現在幾號，幫我查今天發生什麼新聞")], thread_id="test-5")
-    )
-    print(type(messages), messages)
-    print(type(used_tools), used_tools)
-
-    async def main():
-        result = await agent.stream([HumanMessage(content="請問現在幾點")], thread_id="test")
-        async for r in result:
-            print(type(r), r)
-
-    asyncio.run(main())
